requirements.txt
Flask==1.1.2
pyyaml==5.3.1